feat: factorize common behaviors between read and writes
